+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - Lazily CharacterSet only once in SentryBaggageSerialization (#5871)
 - Structured Logging: Log `SentrySDK.logger` calls to `SentrySDKLog` (#5991)
 - The build type in the app context now differentiates between `enterprise` and `adhoc` (#6044)
+- visionOS no longer needs swift's interoperability mode (#6077)
 - Ensure IP address is only inferred by Relay if sendDefaultPii is true (#5877)
 
 ## 8.55.1